chore: setup tracing in Sentry
