+++ apps/backend/src/config/index.ts
@@ -3,7 +3,7 @@ import { fileURLToPath } from "node:url";
 import convict from "convict";
 import dotenv from "dotenv";
 
-import { loadDatabaseConfigFromURL } from "./database";
+import { loadDatabaseConfigFromURL } from "./database-url";
 
 const __dirname = fileURLToPath(new URL(".", import.meta.url));
