+++ apps/backend/src/github/index.ts
@@ -3,6 +3,7 @@ import type { components } from "@octokit/openapi-types";
 export * from "./auth";
 export * from "./client";
 export * from "./comment";
+export * from "./error";
 
 /** @public */
 export type GhApiRepository = components["schemas"]["repository"];
