+++ apps/backend/src/sentry/index.ts
@@ -1,14 +1,16 @@
 import * as Sentry from "@sentry/node";
 
 import config from "@/config";
-import { getOctokitErrorStatus } from "@/github";
+import { getOctokitErrorStatus } from "@/github/error";
 import { isHttp2GoAwayCode0Error } from "@/util/error";
 
 export function setup() {
+  console.log(config.get("sentry.serverDsn"));
   Sentry.init({
     dsn: config.get("sentry.serverDsn"),
     environment: config.get("sentry.environment"),
     release: config.get("releaseVersion"),
+    tracesSampleRate: 1,
     beforeSend(event, hint) {
       const error = hint.originalException;
       // Detect HTTP-like errors
